backend/requirements.txt
google-generativeai==0.3.1
httpx==0.24.1
python-dotenv==1.0.0
tiktoken>=0.3.0
PyPDF2==3.0.1
python-docx==0.8.11
